backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.1
mistralai>=0.0.7

# ML & Data Science
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0
joblib>=1.3.0
scipy>=1.11.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
